% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_poly.R
\name{gen_shape}
\alias{gen_shape}
\title{generate a new shape}
\usage{
gen_shape(nsides = 16, area = 0.3, rescale = TRUE, shape = 0.8)
}
\arguments{
\item{nsides}{the number of sides}

\item{area}{the area of the shape}

\item{rescale}{whether to rescale the shape to a fixed area given by the `area` parameter.}
}
\description{
generate a new shape
}
