% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_poly.R
\name{jitter_poly}
\alias{jitter_poly}
\title{create a new shape by jittering the radii of an existing shape.}
\usage{
jitter_poly(x, sd = 0.003, area = 0.3, rescale = FALSE)
}
\arguments{
\item{x}{the shape to jitter}

\item{sd}{the standard deviation of the noise}

\item{rescale}{whether to rescale the shape to maintain the current area}
}
\description{
create a new shape by jittering the radii of an existing shape.
}
